Refactor backend into modules
