+++ backend/__init__.py
@@ -1,0 +1,14 @@
+from .app import app, run, SessionLocal, CATEGORIES_JSON, load_categories_json, save_categories_json
+from .models import init_db
+from .routes import compute_dashboard_averages
+
+__all__ = [
+    'app',
+    'run',
+    'SessionLocal',
+    'CATEGORIES_JSON',
+    'load_categories_json',
+    'save_categories_json',
+    'init_db',
+    'compute_dashboard_averages',
+]
