+++ run.py
@@ -1,4 +1,4 @@
-from backend.app import run
+from backend import run
 
 if __name__ == '__main__':
     run()
