+++ tests/test_dashboard_helper.py
@@ -4,7 +4,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
 from backend import models
-from backend import app as app_module
+import backend as app_module
 
 class FixedDate(datetime.datetime):
     @classmethod
